FS: extract utils/gently-rm and utils/link to a separate file for future package extraction.
