+++ lib/link.js
@@ -122,7 +122,7 @@ function linkInstall (pkgs, cb) {
             log.verbose('link', 'symlinking %s to %s', pp, target)
             cb()
           } ],
-          [symlink, pp, target],
+          [symlink, pp, target, false, false],
           // do not run any scripts
           rp && [build, [target], npm.config.get('global'), build._noLC, true],
           [resultPrinter, pkg, pp, target, rp]
