+++ lib/utils/bin-links.js
@@ -2,7 +2,7 @@
 
 const fs = require('graceful-fs')
 const path = require('path')
-const linkIfExists = require('./link.js').ifExists
+const linkIfExists = require('./gentle-fs.js').linkIfExists
 const cmdShimIfExists = require('cmd-shim').ifExists
 const asyncMap = require('slide').asyncMap
 const BB = require('bluebird')
@@ -87,11 +87,15 @@ module.exports = function (pkg, folder, global, opts, cb) {
   )
 }
 
+function getLinkOpts (opts, gently) {
+  return Object.assign({}, opts, { gently: gently })
+}
+
 function linkBins (pkg, folder, parent, gtop, opts, cb) {
   if (!pkg.bin || !gtop && path.basename(parent) !== 'node_modules') {
     return cb()
   }
-  var gently = gtop && folder
+  var linkOpts = getLinkOpts(opts, gtop && folder)
   var execMode = parseInt('0777', 8) & (~opts.umask)
   var binRoot = gtop ? opts.globalBin
                      : path.resolve(parent, '.bin')
@@ -101,7 +105,7 @@ function linkBins (pkg, folder, parent, gtop, opts, cb) {
     var dest = path.resolve(binRoot, bin)
     var src = path.resolve(folder, pkg.bin[bin])
 
-    linkBin(src, dest, gently, function (er) {
+    linkBin(src, dest, linkOpts, function (er) {
       if (er) return cb(er)
       // bins should always be executable.
       // XXX skip chmod on windows?
@@ -131,9 +135,9 @@ function linkBins (pkg, folder, parent, gtop, opts, cb) {
   }, cb)
 }
 
-function linkBin (from, to, gently, cb) {
+function linkBin (from, to, opts, cb) {
   if (process.platform !== 'win32') {
-    return linkIfExists(from, to, gently, cb)
+    return linkIfExists(from, to, opts, cb)
   } else {
     return cmdShimIfExists(from, to, cb)
   }
@@ -173,6 +177,6 @@ function linkMans (pkg, folder, parent, gtop, opts, cb) {
     var manSrc = path.resolve(folder, man)
     var manDest = path.join(manRoot, 'man' + sxn, bn)
 
-    linkIfExists(manSrc, manDest, gtop && folder, cb)
+    linkIfExists(manSrc, manDest, getLinkOpts(opts, gtop && folder), cb)
   }, cb)
 }
